requirements.txt
# ============================================
# AI CHATBOT LMS - PYTHON DEPENDENCIES
# Updated: 2024-12-15
# ============================================

# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Vector Database
qdrant-client==1.6.9

# AI & NLP (Updated versions)
openai==1.3.7
langdetect==1.0.9
sentence-transformers==3.3.1
transformers==4.47.1
huggingface_hub==0.26.5

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
numpy==1.26.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
